Simplify sign.retracting assignment in updateSigns.
Motivation for this change was to simplify code, but it also improves
the interaction a little bit. Previously if you clicked a sign while it
was retracting it would continue retracting and not start moving into
place until the retraction finished. Now if you click a sign that is
retracting it stops retracting and starts moving into place right away.

git-svn-id: https://russ.yanofsky.org/svn/site/trunk@103 af7d400b-3b05-0410-8da3-a4763f9e3e29